requirements.txt
anyio==2.2.0
appdirs==1.4.4
argcomplete==1.12.0
argh==0.26.1
argon2-cffi==20.1.0
asgiref==3.2.7
asn1crypto==1.4.0
async-generator==1.10
attrs==20.3.0
Babel==2.9.0
backcall==0.2.0
bcrypt==3.1.7
Beaker==1.10.0
beautifulsoup4==4.9.1
bleach==3.3.0
blivet==3.3.3
blivet-gui==2.2.1
Brlapi==0.8.2
cached-property==1.5.2
cairocffi==1.2.0
CairoSVG==2.5.2
certifi==2020.12.5
cffi==1.14.5
cfgv==2.0.1
chardet==3.0.4
chrome-gnome-shell==0.0.0
click==7.1.2
configobj==5.0.6
coverage==5.5
cryptography==3.4.6
cssselect==0.9.2
cssselect2==0.3.0
cupshelpers==1.0
cycler==0.10.0
dasbus==1.4
dbus-python==1.2.16
decorator==4.4.2
defusedxml==0.7.0rc2
distlib==0.3.1
distro==1.5.0
Django==3.0.6
django-bootstrap4==1.1.1
django-credit-cards==0.4.1
djangorestframework==3.12.4
docker==4.4.4
docker-compose==1.28.6
docker-pycreds==0.4.0
dockerpty==0.4.1
docopt==0.6.2
entrypoints==0.3
filelock==3.0.12
fluidity-sm==0.2.0
fros==1.1
gpg==1.15.1
html5lib==1.1
humanize==0.5.1
identify==1.5.14
idna==2.10
initial-setup==0.3.90
invoke==1.5.0
ipykernel==5.5.0
ipython==7.22.0
ipython-genutils==0.2.0
jedi==0.18.0
Jinja2==2.11.3
json5==0.9.5
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyter-packaging==0.7.12
jupyter-server==1.5.1
jupyterlab==3.0.12
jupyterlab-pygments==0.1.2
jupyterlab-server==2.3.0
kiwisolver==1.3.1
langtable==0.0.54
lexicon==1.0.0
libcomps==0.1.15
lorax===devel
lxml==4.6.3
Mako==1.1.4.dev0
MarkupSafe==1.1.1
matplotlib==3.4.2
mistune==0.8.4
nbclassic==0.2.6
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.2
nest-asyncio==1.5.1
nftables==0.1
nodeenv==0.13.6
notebook==6.3.0
notify2==0.3.1
ntplib==0.3.3
numpy==1.20.1
olefile==0.46
onboard==1.4.1
ordered-set==4.0.1
packaging==20.9
pandocfilters==1.4.3
paramiko==2.7.2
parso==0.8.1
Paste==3.5.0
pexpect==4.8.0
pickleshare==0.7.5
pid==2.2.3
Pillow==8.1.2
ply==3.11
powerline-status==2.8.2
pre-commit==2.11.1
productmd==1.32
progressbar2==3.52.1
prometheus-client==0.9.0
prompt-toolkit==3.0.18
psutil==5.8.0
ptyprocess==0.6.0
pwquality==1.4.4
pyasn1==0.4.8
pycairo==1.20.0
pycdlib==1.11.0
pycparser==2.20
pycrypto==2.6.1
pycryptodome==3.9.7
pycups==2.0.1
pycurl==7.43.0.6
pydbus==0.6.0
pyenchant==3.2.0
pygit2==1.5.0
Pygments==2.8.1
PyGObject==3.40.1
pyinotify==0.9.6
pykickstart==3.32
PyMySQL==1.0.2
PyNaCl==1.4.0
pyOpenSSL==20.0.1
pyparsing==2.4.7
pyparted==3.11.7
pyrsistent==0.17.3
pysendfile==2.0.1
PySocks==1.7.1
pystray==0.17.3
python-augeas==0.5.0
python-dateutil==2.8.1
python-dotenv==0.15.0
python-gettext==4.0
python-manatools==0.0.3
python-meh==0.50
python-pam==1.8.4
python-utils==2.4.0
python-xapp==2.0.2
python-xlib==0.29
pytz==2020.1
pyudev==0.22.0
pyxdg==0.27
PyYAML==5.4.1
pyzmq==22.0.3
requests==2.25.1
requests-file==1.5.1
requests-ftp==0.3.1
rpm==4.16.1.3
selinux==3.2
Send2Trash==1.5.0
sepolicy==3.2
setools==4.4.0
setproctitle==1.1.10
setroubleshoot==1.1
simpleline==1.8
six==1.15.0
slip==0.6.4
slip.dbus==0.6.4
sniffio==1.2.0
sortedcontainers==2.2.2
sos==4.1
soupsieve==2.0.1
sqlparse==0.3.1
swtpm-localca==0.5.2
swtpm-setup==0.5.2
systemd-python==234
Tempita==0.5.1
terminado==0.9.3
terminator==2.1.1
testpath==0.4.4
texttable==1.6.2
tinycss==0.3
tinycss2==1.1.0
toml==0.10.2
tornado==6.1
traitlets==5.0.5
urllib3==1.25.10
virtualenv==20.4.0
voila==0.2.7
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==0.57.0
xcffib==0.9.0
